+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Todo
+- update parse for style string that drops empty style of '-'
+
+## [1.1.0] - 2019-12-05
+### Added
+- convert module functions key2id and id2key
 
 ## [1.0.2] - 2019-10-31
 ### Added
@@ -16,5 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - documentation
 
 
-[Unreleased]: https://github.com/sutton-signwriting/core/compare/v1.0.2...HEAD
+[Unreleased]: https://github.com/sutton-signwriting/core/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/sutton-signwriting/core/compare/v1.1.0...v1.0.2
 [1.0.2]: https://github.com/sutton-signwriting/core/releases/tag/v1.0.2